Port `load_gcs` and `save_gcs` to TQL2
