+++ changelog/next/features/5054--gcs.md
@@ -1,0 +1,2 @@
+The `load_gcs` and `save_gcs` operators are now available in TQL2 to interact
+with Google Cloud Storage.
